Conflicts:
	src/dist/replication/client_lib/replication_common.cpp
	src/dist/replication/lib/replica_context.h
	src/dist/replication/lib/replica_learn.cpp

+++ include/dsn/cpp/serialization_helper/thrift_helper.h
@@ -542,7 +542,7 @@ namespace dsn {
         if (binary_proto != nullptr)
         {
             //the protocol is binary protocol
-            return binary_proto->writeString<char_ptr>(char_ptr(name, strlen(name)));
+            return binary_proto->writeString<char_ptr>(char_ptr(name, static_cast<int>(strlen(name))));
         }
         else
         {
@@ -579,7 +579,7 @@ namespace dsn {
         if (binary_proto != nullptr)
         {
             //the protocol is binary protocol
-            return binary_proto->writeString<char_ptr>(char_ptr(name, strlen(name)));
+            return binary_proto->writeString<char_ptr>(char_ptr(name, static_cast<int>(strlen(name))));
         }
         else
         {
